Re-added the CameraPreview widget
