+++ packages/camera/camera/example/lib/main.dart
@@ -131,7 +131,7 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
     } else {
       return AspectRatio(
         aspectRatio: controller.value.aspectRatio,
-        child: controller.buildPreview(),
+        child: CameraPreview(controller),
       );
     }
   }
