+++ packages/camera/camera/lib/src/camera_controller.dart
@@ -1,0 +1,470 @@
+// Copyright 2018 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+part of camera;
+
+final MethodChannel _channel = const MethodChannel('plugins.flutter.io/camera');
+
+/// Signature for a callback receiving the a camera image.
+///
+/// This is used by [CameraController.startImageStream].
+// ignore: inference_failure_on_function_return_type
+typedef onLatestImageAvailable = Function(CameraImage image);
+
+/// Completes with a list of available cameras.
+///
+/// May throw a [CameraException].
+Future<List<CameraDescription>> availableCameras() async {
+  return CameraPlatform.instance.availableCameras();
+}
+
+/// The state of a [CameraController].
+class CameraValue {
+  /// Creates a new camera controller state.
+  const CameraValue({
+    this.isInitialized,
+    this.errorDescription,
+    this.previewSize,
+    this.isRecordingVideo,
+    this.isTakingPicture,
+    this.isStreamingImages,
+    bool isRecordingPaused,
+  }) : _isRecordingPaused = isRecordingPaused;
+
+  /// Creates a new camera controller state for an uninitialized controller.
+  const CameraValue.uninitialized()
+      : this(
+          isInitialized: false,
+          isRecordingVideo: false,
+          isTakingPicture: false,
+          isStreamingImages: false,
+          isRecordingPaused: false,
+        );
+
+  /// True after [CameraController.initialize] has completed successfully.
+  final bool isInitialized;
+
+  /// True when a picture capture request has been sent but as not yet returned.
+  final bool isTakingPicture;
+
+  /// True when the camera is recording (not the same as previewing).
+  final bool isRecordingVideo;
+
+  /// True when images from the camera are being streamed.
+  final bool isStreamingImages;
+
+  final bool _isRecordingPaused;
+
+  /// True when camera [isRecordingVideo] and recording is paused.
+  bool get isRecordingPaused => isRecordingVideo && _isRecordingPaused;
+
+  /// Description of an error state.
+  ///
+  /// This is null while the controller is not in an error state.
+  /// When [hasError] is true this contains the error description.
+  final String errorDescription;
+
+  /// The size of the preview in pixels.
+  ///
+  /// Is `null` until  [isInitialized] is `true`.
+  final Size previewSize;
+
+  /// Convenience getter for `previewSize.height / previewSize.width`.
+  ///
+  /// Can only be called when [initialize] is done.
+  double get aspectRatio => previewSize.height / previewSize.width;
+
+  /// Whether the controller is in an error state.
+  ///
+  /// When true [errorDescription] describes the error.
+  bool get hasError => errorDescription != null;
+
+  /// Creates a modified copy of the object.
+  ///
+  /// Explicitly specified fields get the specified value, all other fields get
+  /// the same value of the current object.
+  CameraValue copyWith({
+    bool isInitialized,
+    bool isRecordingVideo,
+    bool isTakingPicture,
+    bool isStreamingImages,
+    String errorDescription,
+    Size previewSize,
+    bool isRecordingPaused,
+  }) {
+    return CameraValue(
+      isInitialized: isInitialized ?? this.isInitialized,
+      errorDescription: errorDescription,
+      previewSize: previewSize ?? this.previewSize,
+      isRecordingVideo: isRecordingVideo ?? this.isRecordingVideo,
+      isTakingPicture: isTakingPicture ?? this.isTakingPicture,
+      isStreamingImages: isStreamingImages ?? this.isStreamingImages,
+      isRecordingPaused: isRecordingPaused ?? _isRecordingPaused,
+    );
+  }
+
+  @override
+  String toString() {
+    return '$runtimeType('
+        'isRecordingVideo: $isRecordingVideo, '
+        'isInitialized: $isInitialized, '
+        'errorDescription: $errorDescription, '
+        'previewSize: $previewSize, '
+        'isStreamingImages: $isStreamingImages)';
+  }
+}
+
+/// Controls a device camera.
+///
+/// Use [availableCameras] to get a list of available cameras.
+///
+/// Before using a [CameraController] a call to [initialize] must complete.
+///
+/// To show the camera preview on the screen use a [CameraPreview] widget.
+class CameraController extends ValueNotifier<CameraValue> {
+  /// Creates a new camera controller in an uninitialized state.
+  CameraController(
+    this.description,
+    this.resolutionPreset, {
+    this.enableAudio = true,
+  }) : super(const CameraValue.uninitialized());
+
+  /// The properties of the camera device controlled by this controller.
+  final CameraDescription description;
+
+  /// The resolution this controller is targeting.
+  ///
+  /// This resolution preset is not guaranteed to be available on the device,
+  /// if unavailable a lower resolution will be used.
+  ///
+  /// See also: [ResolutionPreset].
+  final ResolutionPreset resolutionPreset;
+
+  /// Whether to include audio when recording a video.
+  final bool enableAudio;
+
+  int _cameraId;
+  bool _isDisposed = false;
+  StreamSubscription<dynamic> _imageStreamSubscription;
+  Completer<void> _creatingCompleter;
+
+  /// Checks whether [CameraController.dispose] has completed successfully.
+  ///
+  /// This is a no-op when asserts are disabled.
+  void debugCheckIsDisposed() {
+    assert(_isDisposed);
+  }
+
+  /// Initializes the camera on the device.
+  ///
+  /// Throws a [CameraException] if the initialization fails.
+  Future<void> initialize() async {
+    if (_isDisposed) {
+      throw CameraException(
+        'Disposed CameraController',
+        'initialize was called on a disposed CameraController',
+      );
+    }
+    try {
+      _creatingCompleter = Completer<void>();
+      Completer _resolutionCompleter = Completer<Size>();
+
+      _cameraId = await CameraPlatform.instance.createCamera(
+        description,
+        resolutionPreset,
+        enableAudio: enableAudio,
+      );
+
+      unawaited(
+        CameraPlatform.instance
+            .onCameraInitialized(_cameraId)
+            .map((event) {
+              return Size(
+                event.previewWidth,
+                event.previewHeight,
+              );
+            })
+            .first
+            .then((previewSize) => _resolutionCompleter.complete(previewSize)),
+      );
+
+      await CameraPlatform.instance.initializeCamera(_cameraId);
+
+      value = value.copyWith(
+        isInitialized: true,
+        previewSize: await _resolutionCompleter.future,
+      );
+    } on PlatformException catch (e) {
+      throw CameraException(e.code, e.message);
+    }
+    _creatingCompleter.complete();
+    return _creatingCompleter.future;
+  }
+
+  /// Prepare the capture session for video recording.
+  ///
+  /// Use of this method is optional, but it may be called for performance
+  /// reasons on iOS.
+  ///
+  /// Preparing audio can cause a minor delay in the CameraPreview view on iOS.
+  /// If video recording is intended, calling this early eliminates this delay
+  /// that would otherwise be experienced when video recording is started.
+  /// This operation is a no-op on Android.
+  ///
+  /// Throws a [CameraException] if the prepare fails.
+  Future<void> prepareForVideoRecording() async {
+    await CameraPlatform.instance.prepareForVideoRecording();
+  }
+
+  /// Captures an image and saves it to [path].
+  ///
+  /// A path can for example be obtained using
+  /// [path_provider](https://pub.dartlang.org/packages/path_provider).
+  ///
+  /// If a file already exists at the provided path an error will be thrown.
+  /// The file can be read as this function returns.
+  ///
+  /// Throws a [CameraException] if the capture fails.
+  Future<XFile> takePicture() async {
+    if (!value.isInitialized || _isDisposed) {
+      throw CameraException(
+        'Uninitialized CameraController.',
+        'takePicture was called on uninitialized CameraController',
+      );
+    }
+    if (value.isTakingPicture) {
+      throw CameraException(
+        'Previous capture has not returned yet.',
+        'takePicture was called before the previous capture returned.',
+      );
+    }
+    try {
+      value = value.copyWith(isTakingPicture: true);
+      XFile file = await CameraPlatform.instance.takePicture(_cameraId);
+      value = value.copyWith(isTakingPicture: false);
+      return file;
+    } on PlatformException catch (e) {
+      value = value.copyWith(isTakingPicture: false);
+      throw CameraException(e.code, e.message);
+    }
+  }
+
+  /// Start streaming images from platform camera.
+  ///
+  /// Settings for capturing images on iOS and Android is set to always use the
+  /// latest image available from the camera and will drop all other images.
+  ///
+  /// When running continuously with [CameraPreview] widget, this function runs
+  /// best with [ResolutionPreset.low]. Running on [ResolutionPreset.high] can
+  /// have significant frame rate drops for [CameraPreview] on lower end
+  /// devices.
+  ///
+  /// Throws a [CameraException] if image streaming or video recording has
+  /// already started.
+  // TODO(bmparr): Add settings for resolution and fps.
+  Future<void> startImageStream(onLatestImageAvailable onAvailable) async {
+    if (!value.isInitialized || _isDisposed) {
+      throw CameraException(
+        'Uninitialized CameraController',
+        'startImageStream was called on uninitialized CameraController.',
+      );
+    }
+    if (value.isRecordingVideo) {
+      throw CameraException(
+        'A video recording is already started.',
+        'startImageStream was called while a video is being recorded.',
+      );
+    }
+    if (value.isStreamingImages) {
+      throw CameraException(
+        'A camera has started streaming images.',
+        'startImageStream was called while a camera was streaming images.',
+      );
+    }
+
+    try {
+      await _channel.invokeMethod<void>('startImageStream');
+      value = value.copyWith(isStreamingImages: true);
+    } on PlatformException catch (e) {
+      throw CameraException(e.code, e.message);
+    }
+    const EventChannel cameraEventChannel =
+        EventChannel('plugins.flutter.io/camera/imageStream');
+    _imageStreamSubscription =
+        cameraEventChannel.receiveBroadcastStream().listen(
+      (dynamic imageData) {
+        onAvailable(CameraImage.fromPlatformData(imageData));
+      },
+    );
+  }
+
+  /// Stop streaming images from platform camera.
+  ///
+  /// Throws a [CameraException] if image streaming was not started or video
+  /// recording was started.
+  Future<void> stopImageStream() async {
+    if (!value.isInitialized || _isDisposed) {
+      throw CameraException(
+        'Uninitialized CameraController',
+        'stopImageStream was called on uninitialized CameraController.',
+      );
+    }
+    if (value.isRecordingVideo) {
+      throw CameraException(
+        'A video recording is already started.',
+        'stopImageStream was called while a video is being recorded.',
+      );
+    }
+    if (!value.isStreamingImages) {
+      throw CameraException(
+        'No camera is streaming images',
+        'stopImageStream was called when no camera is streaming images.',
+      );
+    }
+
+    try {
+      value = value.copyWith(isStreamingImages: false);
+      await _channel.invokeMethod<void>('stopImageStream');
+    } on PlatformException catch (e) {
+      throw CameraException(e.code, e.message);
+    }
+
+    await _imageStreamSubscription.cancel();
+    _imageStreamSubscription = null;
+  }
+
+  /// Start a video recording.
+  ///
+  /// The video is returned as a [XFile] after calling [stopVideoRecording].
+  /// Throws a [CameraException] if the capture fails.
+  Future<void> startVideoRecording() async {
+    if (!value.isInitialized || _isDisposed) {
+      throw CameraException(
+        'Uninitialized CameraController',
+        'startVideoRecording was called on uninitialized CameraController',
+      );
+    }
+    if (value.isRecordingVideo) {
+      throw CameraException(
+        'A video recording is already started.',
+        'startVideoRecording was called when a recording is already started.',
+      );
+    }
+    if (value.isStreamingImages) {
+      throw CameraException(
+        'A camera has started streaming images.',
+        'startVideoRecording was called while a camera was streaming images.',
+      );
+    }
+
+    try {
+      await CameraPlatform.instance.startVideoRecording(_cameraId);
+      value = value.copyWith(isRecordingVideo: true, isRecordingPaused: false);
+    } on PlatformException catch (e) {
+      throw CameraException(e.code, e.message);
+    }
+  }
+
+  /// Stops the video recording and returns the file where it was saved.
+  ///
+  /// Throws a [CameraException] if the capture failed.
+  Future<XFile> stopVideoRecording() async {
+    if (!value.isInitialized || _isDisposed) {
+      throw CameraException(
+        'Uninitialized CameraController',
+        'stopVideoRecording was called on uninitialized CameraController',
+      );
+    }
+    if (!value.isRecordingVideo) {
+      throw CameraException(
+        'No video is recording',
+        'stopVideoRecording was called when no video is recording.',
+      );
+    }
+    try {
+      XFile file = await CameraPlatform.instance.stopVideoRecording(_cameraId);
+      value = value.copyWith(isRecordingVideo: false);
+      return file;
+    } on PlatformException catch (e) {
+      throw CameraException(e.code, e.message);
+    }
+  }
+
+  /// Pause video recording.
+  ///
+  /// This feature is only available on iOS and Android sdk 24+.
+  Future<void> pauseVideoRecording() async {
+    if (!value.isInitialized || _isDisposed) {
+      throw CameraException(
+        'Uninitialized CameraController',
+        'pauseVideoRecording was called on uninitialized CameraController',
+      );
+    }
+    if (!value.isRecordingVideo) {
+      throw CameraException(
+        'No video is recording',
+        'pauseVideoRecording was called when no video is recording.',
+      );
+    }
+    try {
+      await CameraPlatform.instance.pauseVideoRecording(_cameraId);
+      value = value.copyWith(isRecordingPaused: true);
+    } on PlatformException catch (e) {
+      throw CameraException(e.code, e.message);
+    }
+  }
+
+  /// Resume video recording after pausing.
+  ///
+  /// This feature is only available on iOS and Android sdk 24+.
+  Future<void> resumeVideoRecording() async {
+    if (!value.isInitialized || _isDisposed) {
+      throw CameraException(
+        'Uninitialized CameraController',
+        'resumeVideoRecording was called on uninitialized CameraController',
+      );
+    }
+    if (!value.isRecordingVideo) {
+      throw CameraException(
+        'No video is recording',
+        'resumeVideoRecording was called when no video is recording.',
+      );
+    }
+    try {
+      await CameraPlatform.instance.resumeVideoRecording(_cameraId);
+      value = value.copyWith(isRecordingPaused: false);
+    } on PlatformException catch (e) {
+      throw CameraException(e.code, e.message);
+    }
+  }
+
+  /// Returns a widget showing a live camera preview.
+  Widget buildPreview() {
+    if (!value.isInitialized || _isDisposed) {
+      throw CameraException(
+        'Uninitialized CameraController',
+        'buildView() was called on uninitialized CameraController.',
+      );
+    }
+    try {
+      return CameraPlatform.instance.buildPreview(_cameraId);
+    } on PlatformException catch (e) {
+      throw CameraException(e.code, e.message);
+    }
+  }
+
+  /// Releases the resources of this camera.
+  @override
+  Future<void> dispose() async {
+    if (_isDisposed) {
+      return;
+    }
+    _isDisposed = true;
+    super.dispose();
+    if (_creatingCompleter != null) {
+      await _creatingCompleter.future;
+      await CameraPlatform.instance.dispose(_cameraId);
+    }
+  }
+}
