+++ packages/camera/camera/lib/src/camera_image.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-part of 'camera.dart';
+part of camera;
 
 /// A single color plane of image data.
 ///
